+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Users/ResendTemporaryPassword/Validator.cs
@@ -1,0 +1,12 @@
+using FluentValidation;
+
+namespace TimeSheet.Modules.EmploymentManagement.Application.Users.ResendTemporaryPassword;
+
+public sealed class Validator : AbstractValidator<Command>
+{
+    public Validator()
+    {
+        RuleFor(x => x.UserId)
+            .NotEmpty();
+    }
+}